jwt-fastapi/pyproject.toml
[project]
name = "jwt-fastapi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt[crypto]>=2.12.1",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.42.0",
]
